Implemented Basic Map-to-html converter
The parser can currently take a 2-dimensional array and renders it
as a series of "tile" divs, and returns them all within a single
"map" div. The maps are currently only capable of expressing
altitude, and the return is a simple DOM object. More will need
to be done to improve the map definition and return formats.

This closes #1