Actor extension support added
Actor objects now have an add function that can be used to allow
other objects to enhance the actor object. It calls the addTo
method of the feature object and returns the actor object. The
advantage of this method over just calling addTo is that it allows
chaining. The update loop now loops over an array of functions that
can be added to by feature objects.