Physics with vectors
The physics extensions to the actor object now understand vectors
and make use of them. The motor method not only moves the object,
but is now responsible for adding friction. The jump function
also enables gravity on the actor, along with a sense of a floor
at z = 0. Once the maps are upgraded, the actors should be able to
know the height of the ground and land accordingly.