Replaced Map with Terrain
The terrain object works the same as the map did, but it acts a
bit like an actor. It also includes a constructor for the tiles.
Instead of just generating a bunch of HTML to render the terrain,
the new object creates a bunch of Tile objects on init, and when
asked to render it just adjusts the heights on everything.

Terrain objects have a function for roomEnter, and adjust
to the new room imediately. This allows it to just sit in the
game and do it's thing.

Major step on issue #14