HTML Structure improved
The game is now represented in the HTML as a single div,
everything else is generated on the fly by the javascript. A bit
more thought has been put into what's inside of what. For example,
the terrain and the world no longer share the same div, but one is
inside the other, as it should be.