Code standards changes
Removed the closure around everything in the source files. It
didn't really serve much of a purpose. The idea is to stop
"private" functions from leaking into the global namespace, but
private functions fight the whole philosophy of Aukera. Hiding
the contents of the closure from the global scope also hides them
from debugging tools, and the IDE's outline and code assist tools.

Multi-line var statements were also replaced with multiple var
statemens, as this seems easier to maintain.