Added a common header to all javascript files
Any javascript files that are part of the aukera library should
have the same two lines at the top, a jshint comment, and a line
that activates "use strict" and declares auk if it doesn't exist
already. It is assumed that a yet-to-be-written concatenator can
remove these 2 lines in favour of an all-encompasing closure with
a function-level "use strict".

auk is also no longer assumed to be a child of window, which
should make building for non-browser environments easier.